unified ErrorFunction and NoisyErrorFunction
